Restructure folder by renaming 'hosting-overview-v3' to 'hosting-overview'.
