+++ client/a8c-for-agencies/sections/marketplace/hosting-overview/hooks/use-compact-on-scroll.ts
