+++ client/a8c-for-agencies/sections/marketplace/hosting-overview/hosting-content/standard-agency-hosting/index.tsx
